gitlab: Add issue template for reporting a generic bug
When reporting an issue in gitlab, the project can define a template for
various scenarios which are meant to guide the users to add the relevant
information the project needs to the reported issue.

Add a template for a generic bug report against libvirt. The template
adds sections which motivate users to add version information and also
link to documentation about fetching logs and such.

Another section then motivates the users to describe the steps taken and
also the expected outcome if it's not obvious.

Finally a template also automatically applies one or more labels, so new
issues are already partially pre-triaged.

Note that markdown seems to be the only supported format for now.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Message-Id: <187d93a66f4480adede6374876a2d0fc9a29b3b9.1606801687.git.pkrempa@redhat.com>