conf: Move some of virDomainMemoryDef members into a union
The structure has two sets of members: some for the target and
some for the source. The latter is model dependant (e.g. path to
a nvdimm device applies only to NVDIMM model). Move the members
into an union so that it is obvious which members apply to which
model. This way it's easier to maintain code cleanliness when
introducing a new model.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <57e76dabdaa7f273147ed016de821050c1710ac1.1606998426.git.mprivozn@redhat.com>
Reviewed-by: Daniel Henrique Barboza <danielhb413@gmail.com>
Tested-by: Daniel Henrique Barboza <danielhb413@gmail.com>
Tested-by: Han Han <hhan@redhat.com>