qemu: Build command line for virtio-mem
The majority of the code is prepared since virtio-pmem. The alias
for virtio-mem is new ("virtiomem"). And again, prealloc doesn't
make much sense - the whole point is to dynamically allocate
memory for guest and possibly return it to the host (e.g. if the
host is running low on memory) and this wouldn't happen with
prealloc.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <ec62c290f42a9fc8256e8b4fa64b6c093e818abd.1606998426.git.mprivozn@redhat.com>
Reviewed-by: Daniel Henrique Barboza <danielhb413@gmail.com>
Tested-by: Daniel Henrique Barboza <danielhb413@gmail.com>
Tested-by: Han Han <hhan@redhat.com>