domain_conf.c: move vendor, product and tray checks to validate callback
The 'tray' check isn't a XML parse specific code and can be pushed
to the validate callback, in virDomainDiskDefValidate().

'vendor' and 'product' string sizes are already checked by the
domaincommon.rng schema, but can be of use in the validate callback
since not all scenarios will go through the XML parsing.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>
Message-Id: <20201207135435.884594-5-danielhb413@gmail.com>
Reviewed-by: Michal Privoznik <mprivozn@redhat.com>