qemuBlockJobInfoTranslate: Take job type from qemuBlockJobDataPtr
Commit f5e8715a8b4 added logic which adds some fake job info when qemu
didn't return anything but in such case the job type would not be set.

Since we already have the proper job type recorded in qemuBlockJobDataPtr
which the caller fetched, we can use this it and also remove the lookup
from the disk which was necessary prior to the conversion to
qemuBlockJobDataPtr.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Message-Id: <b03dfdd52baeaa16543e39420259dc0b1cd05fce.1607360209.git.pkrempa@redhat.com>
Reviewed-by: Michal Privoznik <mprivozn@redhat.com>