To start multiple KVM guests from one qcow2 image with transient disk option
Hello,

I would like to start multiple KVM guests from one qcow2 image, and
discard the changes which the KVM guests done.

transient disk option is useful for discarding the changes, however,
we cannot start multiple KVM guest from one qcow2 image because the
image is write-locked by the first guest to be started.

I suppose the disk which transient option is enabled don't need to
get the write lock because any changes go to the overlay image, and
the overlay image is removed when the guest shutdown.

qemu has 'locking' option and the write lock is disabled when locking=off.
To implement that, I have two ideas. I would appreciate it if you could
give me the ideas which way is better (or another way).

1. Add an element to handle 'locking' qemu option. Like as:

    <disk type='file' device='disk'>
      <driver name='qemu' type='qcow2' locking='off'/>
      <source file='/var/lib/libvirt/images/guest.qcow2'/>
      <target dev='vda' bus='virtio'/>
      <transient/>
    </disk>

2. Add locking=off internally only if the transient disk option is enabled.
   The sample code is as follows:

Message-Id: <20201212015848.7vzqjilkdmppupik@gabell>