virfile: refactor virFileNBDDeviceAssociate
The only reason why virstoragefile.h needs to be included in virfile.h
is that virFileNBDDeviceAssociate() takes virStorageFileFormat argument.
The function doesn't need the enum value as it converts the value to
string and uses only that.

Change the argument to string which will allow us to remove that
include.

The extra seemingly unrelated include changes is because all of the
added includes where indirectly provided by virfile.h having
virstoragefile.h.

Signed-off-by: Pavel Hrdina <phrdina@redhat.com>
Message-Id: <2469a62f02ecccff769b9e989dcc0d34fdba5799.1607960951.git.phrdina@redhat.com>