vmx: Don't error out on missing filename for cdrom
This is perfectly valid in VMWare and the VM just boots with an empty drive.  We
used to just skip the whole drive before, but since we changed how we parse
empty cdrom drives this now results in an error and the user not being able to
even dump the XML.  Instead of erroring out, just keep the drive empty.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1903953

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>
Message-Id: <fef74080eab825957802e34c0cb20c8146bb1abb.1608551303.git.mkletzan@redhat.com>
Reviewed-by: Tim Wiederhake <twiederh@redhat.com>