nodedev: add ability to list and parse defined mdevs
This adds some internal API to query for persistent mediated devices
that are defined by mdevctl. Following commits will make use of this
information. This just provides the infrastructure and tests for this
feature. One test verifies that we are executing mdevctl with the proper
arguments, and the other test verifies that we can parse the returned
JSON and convert it to our own internal node device representations.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>
Message-Id: <20201224141445.163819-6-jjongsma@redhat.com>
Tested-by: Yan Fu <yafu@redhat.com>
tested-by should not be given.