nodedev: add helper functions to remove node devices
When a mediated device is stopped or undefined by an application outside
of libvirt, we need to remove it from our list of node devices within
libvirt. This patch introduces virNodeDeviceObjListRemoveLocked() and
virNodeDeviceObjListForEach() (which are analogous to other types of
object lists in libvirt) to facilitate that. They will be used in coming
commits.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>
Message-Id: <20201224141445.163819-9-jjongsma@redhat.com>
Tested-by: Yan Fu <yafu@redhat.com>
tested-by should not be given.