nodedev: rename dataReady to udevReady
The udev thread handles received udev events. This is accomplished by
setting dataReady to 'true' and signalling the thread.

We also want to use this thread to handle mdev events, so we'll need to
add another variable to indicate which event has woken the thread. To
prepare for this, rename 'dataReady' so that it is clear that when this
variable is set, the udev event needs to be handled. Mdevctl events will
use a different variable name.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>
Message-Id: <20201224141445.163819-11-jjongsma@redhat.com>
Tested-by: Yan Fu <yafu@redhat.com>
tested-by should not be given.