nodedev: Refresh mdev devices when changes are detected
We need to periodically query mdevctl for changes to device definitions
since an administrator can define new devices with mdevctl outside of
libvirt.

In the future, mdevctl may add a way to signal device add/remove via
events, but for now we resort to a bit of a workaround: monitoring the
mdevctl config directories for changes to files. When a change is
detected, we query mdevctl and update our device list. The mdevctl
querying is handled by the existing udev thread.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>
Message-Id: <20201224141445.163819-12-jjongsma@redhat.com>
Tested-by: Yan Fu <yafu@redhat.com>
tested-by should not be given.